完善合同编辑界面相关接口

+++ CarRentalManagement/common/contract.h
@@ -136,7 +136,7 @@ public:
 
     float   deliverySizes;
     float   structureLevel;
-    float   taxRate;
+    int     taxRate;
 
     QDate   signedDate;
     QDate   beginDate;
